[react-components] Add check-api gating + rename primitives export to internal
- Wire up API Extractor (check-api) on the `experimental` entry point; commit
  the generated etc/react-components.experimental.report.api.md so public API
  changes surface as a committed report diff (enforced by CI's git diff check).
- Rename the `./primitives` subpath export to `./internal` (source file, root
  barrel, package.json exports); repoint the cbac-components + storybook
  consumers and update the package docs. The `internal` entry is intentionally
  not gated by check-api.
- Add a PR workflow that labels PRs `api-change` when the committed report
  changes, removing the label when it no longer does.
- turbo.json: generalize the check-api cache input to api-extractor*.json.
