+++ packages/react-components/CLAUDE.md
@@ -41,7 +41,7 @@ If a skill ever conflicts with this file or `CONTRIBUTING.md`, those win — fla
 ### Reuse before writing
 
 - **Reuse from `src/base-components/` first.** This folder contains shared internal UI primitives (`action-button/`, `checkbox/`, `combobox/`, `dialog/`, `draggable-list/`, `icon/`, `search-bar/`, `searchable-menu/`, `select/`, `skeleton/`, `switch/`, `tooltip/`). Check this folder before writing a new UI primitive — reusing avoids visual / behavioral drift
-- **Public primitives barrel: `src/public/primitives.ts`.** Enumerates primitives that are intentionally re-exported (`ActionButton`, `Dialog`, `SkeletonBar`, `Tooltip`, etc.). Anything in `src/base-components/` not re-exported through `primitives.ts` stays internal-only
+- **Public re-export barrel: `src/public/internal.ts`** (exposed as the `@osdk/react-components/internal` subpath). Enumerates primitives that are intentionally re-exported (`ActionButton`, `Dialog`, `SkeletonBar`, `Tooltip`, etc.). Anything in `src/base-components/` not re-exported through `internal.ts` stays internal-only
 - **NEVER export UI primitives from a component folder.** Anything in `src/base-components/` is internal. If a primitive is reusable across components, move it into `src/base-components/` so the next component can pick it up
 - **Prefer `@base-ui/react` over primitive HTML for interactive elements.** When you need a button, dialog, menu, popover, tooltip, select, switch, or similar interactive primitive, reach for the `@base-ui/react` component (it's a workspace dependency) rather than a raw `<button>`, `<div role="dialog">`, etc. Base-ui handles accessibility, keyboard, and focus correctly. Plain semantic HTML (`<div>`, `<section>`, `<span>`, `<a>`) is fine for non-interactive layout
 - **Prefer Blueprint icons (`@blueprintjs/icons`) for iconography.** Use Blueprint's icon set rather than inlining SVGs, importing from another icon library, or hand-rolling glyphs. Verify the icon name exists at the chosen export path before committing — typos surface at runtime, not build time
