+++ packages/react-components/api-extractor.experimental.json
@@ -1,0 +1,11 @@
+{
+  "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
+  "extends": "@osdk/monorepo.api-extractor/base.json",
+  "mainEntryPointFilePath": "<projectFolder>/build/types/public/experimental.d.ts",
+  "apiReport": {
+    "reportFileName": "react-components.experimental.report.api.md"
+  },
+  "docModel": {
+    "apiJsonFilePath": "<projectFolder>/build/api-extractor/react-components.experimental.api.json"
+  }
+}
