+++ turbo.json
@@ -76,7 +76,7 @@
 
     "check-api": {
       "outputLogs": "new-only",
-      "inputs": ["api-extractor.json", "build/esm/**/*", "api-extractor.json"],
+      "inputs": ["api-extractor*.json", "build/esm/**/*"],
       "outputs": ["etc/*.report.api.md", "build/api-extractor/**/*"],
       "dependsOn": ["transpileTypes", "@osdk/monorepo.api-extractor#check-api"]
     },
